+++ HISTORY.md
@@ -1,6 +1,6 @@
 ## v0.2.24 add category command
 
--   category use as: ```fofa category -i input.csv -c Category [-o category.csv]```
+-   category use as: ```fofa category -i input.csv [-o category.csv -rFile relation.csv -sField source -tField target1,target2...]```
 
 ## v0.2.23 add headline in search command